re-added old functions so plex create doesn't break for now
