--- internal/ipwl/generate_io.go
@@ -1,157 +1,0 @@
-package ipwl
-
-import (
-	"fmt"
-	"io/ioutil"
-	"log"
-	"os"
-	"path/filepath"
-
-	"github.com/labdao/plex/internal/ipfs"
-	"github.com/labdao/plex/internal/web3"
-)
-
-func findMatchingFiles(inputDir string, tool Tool, layers int) (map[string][]string, error) {
-	inputFilepaths := make(map[string][]string)
-
-	err := searchMatchingFiles(inputDir, tool, layers, 0, inputFilepaths)
-	if err != nil {
-		return nil, err
-	}
-
-	return inputFilepaths, nil
-}
-
-func searchMatchingFiles(inputDir string, tool Tool, layers int, currentLayer int, inputFilepaths map[string][]string) error {
-	if currentLayer > layers {
-		return nil
-	}
-
-	for inputName, input := range tool.Inputs {
-		if input.Type == "File" {
-			for _, globPattern := range input.Glob {
-				matches, err := filepath.Glob(filepath.Join(inputDir, globPattern))
-				if err != nil {
-					return err
-				}
-
-				inputFilepaths[inputName] = append(inputFilepaths[inputName], matches...)
-			}
-		}
-	}
-
-	// Search subdirectories
-	subDirs, err := ioutil.ReadDir(inputDir)
-	if err != nil {
-		return err
-	}
-
-	for _, subDir := range subDirs {
-		if subDir.IsDir() {
-			err := searchMatchingFiles(filepath.Join(inputDir, subDir.Name()), tool, layers, currentLayer+1, inputFilepaths)
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-func generateCombinationsRecursive(keys []string, values map[string][]string, index int, combination map[string]string, combinations *[]map[string]string) {
-	if index == len(keys) {
-		newCombination := make(map[string]string)
-		for key, value := range combination {
-			newCombination[key] = value
-		}
-		*combinations = append(*combinations, newCombination)
-		return
-	}
-
-	key := keys[index]
-	for _, value := range values[key] {
-		combination[key] = value
-		generateCombinationsRecursive(keys, values, index+1, combination, combinations)
-	}
-}
-
-func generateInputCombinations(inputFilepaths map[string][]string) []map[string]string {
-	var combinations []map[string]string
-	keys := make([]string, 0, len(inputFilepaths))
-
-	for key := range inputFilepaths {
-		keys = append(keys, key)
-	}
-
-	generateCombinationsRecursive(keys, inputFilepaths, 0, make(map[string]string), &combinations)
-
-	return combinations
-}
-
-func createIOEntries(toolInfo ToolInfo, tool Tool, inputCombinations []map[string]string) []IO {
-	var ioData []IO
-
-	var userID string
-
-	if web3.IsValidEthereumAddress(os.Getenv("RECIPIENT_WALLET")) {
-		userID = os.Getenv("RECIPIENT_WALLET")
-	} else {
-		fmt.Println("RECIPIENT_WALLET is not a valid Ethereum address. Using empty string as user ID.")
-		userID = ""
-	}
-
-	for _, combination := range inputCombinations {
-		ioEntry := IO{
-			Tool:    toolInfo,
-			State:   "created",
-			Inputs:  map[string]FileInput{},
-			Outputs: map[string]Output{},
-			UserID:  userID,
-		}
-
-		for inputName, path := range combination {
-			_, fileName := filepath.Split(path)
-
-			cid, err := ipfs.WrapAndPinFile(path)
-			if err != nil {
-				log.Printf("Error getting CID for file %s: %v", path, err)
-				continue
-			}
-
-			ioEntry.Inputs[inputName] = FileInput{
-				Class:    "File",
-				FilePath: fileName,
-				IPFS:     cid,
-			}
-		}
-
-		for outputName, output := range tool.Outputs {
-			if output.Type == "File" {
-				ioEntry.Outputs[outputName] = FileOutput{
-					Class: "File",
-				}
-			} else if output.Type == "Array" && output.Item == "File" {
-				ioEntry.Outputs[outputName] = ArrayFileOutput{
-					Class: "Array",
-					Files: []FileOutput{},
-				}
-			}
-		}
-
-		ioData = append(ioData, ioEntry)
-	}
-
-	return ioData
-}
-
-func CreateIOJson(inputDir string, tool Tool, toolInfo ToolInfo, layers int) ([]IO, error) {
-	inputFilepaths, err := findMatchingFiles(inputDir, tool, layers)
-	if err != nil {
-		return nil, err
-	}
-
-	inputCombinations := generateInputCombinations(inputFilepaths)
-	ioData := createIOEntries(toolInfo, tool, inputCombinations)
-
-	return ioData, nil
-}
